refactor: Move OSM config parameter to the osm module
