+++ src/main/java/org/opentripplanner/graph_builder/module/osm/parameters/OsmDefaultsConfig.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.standalone.config.feed;
+package org.opentripplanner.graph_builder.module.osm.parameters;
 
 import static org.opentripplanner.standalone.config.framework.json.OtpVersion.NA;
