+++ src/main/java/org/opentripplanner/graph_builder/module/osm/parameters/OsmExtractConfig.java
@@ -1,10 +1,11 @@
-package org.opentripplanner.standalone.config.feed;
+package org.opentripplanner.graph_builder.module.osm.parameters;
 
 import java.net.URI;
 import java.time.ZoneId;
 import java.util.Objects;
 import java.util.Optional;
 import org.opentripplanner.graph_builder.module.osm.WayPropertySetSource;
+import org.opentripplanner.standalone.config.feed.DataSourceConfig;
 
 /**
  * Configure an OpenStreetMap extract.
