+++ src/main/java/org/opentripplanner/graph_builder/module/osm/parameters/OsmExtractConfigBuilder.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.standalone.config.feed;
+package org.opentripplanner.graph_builder.module.osm.parameters;
 
 import java.net.URI;
 import java.time.ZoneId;
