+++ src/main/java/org/opentripplanner/openstreetmap/OpenStreetMapProvider.java
@@ -13,10 +13,10 @@ import org.opentripplanner.graph_builder.ConfiguredDataSource;
 import org.opentripplanner.graph_builder.module.osm.OSMDatabase;
 import org.opentripplanner.graph_builder.module.osm.WayPropertySet;
 import org.opentripplanner.graph_builder.module.osm.WayPropertySetSource;
+import org.opentripplanner.graph_builder.module.osm.parameters.OsmDefaultsConfig;
+import org.opentripplanner.graph_builder.module.osm.parameters.OsmExtractConfig;
+import org.opentripplanner.graph_builder.module.osm.parameters.OsmExtractConfigBuilder;
 import org.opentripplanner.openstreetmap.model.OSMProvider;
-import org.opentripplanner.standalone.config.feed.OsmDefaultsConfig;
-import org.opentripplanner.standalone.config.feed.OsmExtractConfig;
-import org.opentripplanner.standalone.config.feed.OsmExtractConfigBuilder;
 import org.opentripplanner.util.lang.ToStringBuilder;
 import org.opentripplanner.util.logging.ProgressTracker;
 import org.slf4j.Logger;
